create-test: be more careful about name cases.

+++ make/create-test.sh
@@ -2,14 +2,6 @@
 
 # Create a unit test `test-foo.h`.
 
-testname="test-$1.h"
-
-if [ -f "$testname" ]
-then
-    echo "Can't create $testname: file exists" >&2
-    exit 1
-fi
-
 kebab="$(python3 -c "print('$1'.lower().replace('_', '-'))")"
 snake="$(python3 -c "print('$kebab'.lower().replace('-', '_'))")"
 camel="$(python3 -c "print('$kebab'.title().replace('-', ''))")"
@@ -17,6 +9,14 @@ camel="$(python3 -c "print('$kebab'.title().replace('-', ''))")"
 echo "kebab = $kebab"
 echo "snake = $snake"
 echo "camel = $camel"
+
+testname="test-${kebab}.h"
+
+if [ -f "$testname" ]
+then
+    echo "Can't create $testname: file exists" >&2
+    exit 1
+fi
 
 cat > "$testname" <<EOF
 #include "$kebab.h"
@@ -29,7 +29,7 @@ public:
 
     void test_instantiate()
     {
-        (void)${snake}();
+        (void)${1}();
     }
 
 };
